Adds save and load feature to the game.
Adds save and load feature to the game. Now you can save your badges and
inventory items to local storage. The item will be `stringfy`ied to
create an inline JSON string that will be stored in the local storage.
The stored string data will be then parsed and returned and then
inserted into the player's badges and inventory objects.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>